Add #clock_in, #clock_out, and #index endpoints
Adding the logic for users to record the beginning and end of their
sleep session and list them all.

Since a user will only have a single active sleep session at any given
time we won't need to pass the id, so we are setting the patch endpoint
on a collection instead of on a member.